Moved config file and added more to comment
