+++ src/components/messageListener.ts
@@ -3,7 +3,7 @@ import { applyBonusByUserId } from './coin';
 import { client } from '../bot';
 import { TextChannel } from 'discord.js';
 import { logError } from './logger';
-import { vars } from '../../config/config';
+import { vars } from '../config';
 
 const PC_CATEGORY_ID: string = vars.PC_CATEGORY_ID;
 const HONEYPOT_CHANNEL_ID: string = vars.HONEYPOT_CHANNEL_ID;
@@ -28,6 +28,7 @@ const detectSpammersAndTrollsNotByHoneypot = (message: Message): boolean => {
 
 /*
  * Punish spammers/trolls/people who got hacked
+ * Return true if someone of this kind is detected, false otherwise
  */
 const punishSpammersAndTrolls = async (message: Message): Promise<boolean> => {
   if (detectSpammersAndTrollsNotByHoneypot(message) || message.channel.id === HONEYPOT_CHANNEL_ID) {
